Support a crossorigin attribute in <link rel=webbundle>
Support a crossorigin attribute in <link rel=webbundle> so that UA can
send a credential to the server.

Spec: https://html.spec.whatwg.org/multipage/#cors-settings-attribute

Since a webbundle is a new feature, a request's mode is always CORS for
a webbundle. A crossorigin attribute only affects how UA sends a
credential to the server.

A general behavior of a crossorigin attribute is well tested in existing
tests, such as fast/dom/HTMLLinkElement/link-crossOrigin.htm.

This CL tests only the effects of a crossorigin attribute used in
<link rel=webbundle>.

BUG=1149816

Change-Id: I354e203a3a536bc829ef7ba6a97d4edd6a6340a0